Move mplFigViewer to the widgets module
