+++ gwhat/widgets/mplfigureviewer.py
